annotation: author :has_many books
i don't wanna bitch too much. in the end it's your book, but it is
not very motivating for the reader to read the same examples over
and over again. i highly recommend to either create one big example
that stretches throughout the whole book (see "agile web development
with ruby on rails" by DHH and Dave Thomas for a good example on
that) or if you wanna stick with small simple examples at least try
to choose the most natural ones and leave out the setup steps for
all those rails apps.

please don't hate me for being that open with you. i really just
want to help improving your book as much as possible.